requirements.txt
torch
torchvision
timm
gradio
bing-image-downloader
Pillow
scikit-learn
numpy
pandas
matplotlib
